Simplify order book
