+++ x/ibcdex/types/buy_order_book.go
@@ -1,0 +1,92 @@
+ // x/ibcdex/types/buy_order_book.go
+package types
+
+ // AppendOrder appends an order in buy order book
+ func (b *BuyOrderBook) AppendOrder(creator string, amount int32, price int32) (int32, error) {
+	 return b.Book.appendOrder(creator, amount, price, Increasing)
+ }
+
+// FillSellOrder try to fill the sell order with the order book and returns all the side effects
+func (b *BuyOrderBook) FillSellOrder(order Order) (
+	remainingSellOrder Order,
+	liquidated []Order,
+	gain int32,
+	filled bool,
+) {
+	var liquidatedList []Order
+	totalGain := int32(0)
+	remainingSellOrder = order
+
+	// Liquidate as long as there is match
+	for {
+		var match bool
+		var liquidation Order
+		remainingSellOrder, liquidation, gain, match, filled = b.liquidateFromSellOrder(
+			remainingSellOrder,
+		)
+		if !match {
+			break
+		}
+
+		// Update gains
+		totalGain += gain
+
+		// Update liquidated
+		liquidatedList = append(liquidatedList, liquidation)
+
+		if filled {
+			break
+		}
+	}
+
+	return remainingSellOrder, liquidatedList, totalGain, filled
+}
+
+ // liquidateFromSellOrder liquidates the first buy order of the book from the sell order
+ // if no match is found, return false for match
+ func (b *BuyOrderBook) liquidateFromSellOrder(order Order) (
+	 remainingSellOrder Order,
+	 liquidatedBuyOrder Order,
+	 gain int32,
+	 match bool,
+	 filled bool,
+ ) {
+	 remainingSellOrder = order
+
+	 // No match if no order
+	 orderCount := len(b.Book.Orders)
+	 if orderCount == 0 {
+		 return order, liquidatedBuyOrder, gain, false, false
+	 }
+
+	 // Check if match
+	 highestBid := b.Book.Orders[orderCount-1]
+	 if order.Price > highestBid.Price {
+		 return order, liquidatedBuyOrder, gain, false, false
+	 }
+
+	 liquidatedBuyOrder = *highestBid
+
+	 // Check if sell order can be entirely filled
+	 if highestBid.Amount >= order.Amount {
+		 remainingSellOrder.Amount = 0
+		 liquidatedBuyOrder.Amount = order.Amount
+		 gain = order.Amount * highestBid.Price
+
+		 // Remove highest bid if it has been entirely liquidated
+		 highestBid.Amount -= order.Amount
+		 if highestBid.Amount == 0 {
+			 b.Book.Orders = b.Book.Orders[:orderCount-1]
+		 } else {
+			 b.Book.Orders[orderCount-1] = highestBid
+		 }
+		 return remainingSellOrder, liquidatedBuyOrder, gain, true, true
+	 }
+
+	 // Not entirely filled
+	 gain = highestBid.Amount * highestBid.Price
+	 b.Book.Orders = b.Book.Orders[:orderCount-1]
+	 remainingSellOrder.Amount -= highestBid.Amount
+
+	 return remainingSellOrder, liquidatedBuyOrder, gain, true, false
+ }
