+++ x/ibcdex/types/sell_order_book.go
@@ -1,0 +1,93 @@
+// x/ibcdex/types/sell_order_book.go
+package types
+
+// AppendOrder appends an order in sell order book
+func (s *SellOrderBook) AppendOrder(creator string, amount int32, price int32) (int32, error) {
+	return s.Book.appendOrder(creator, amount, price, Decreasing)
+}
+
+// FillBuyOrder try to fill the buy order with the order book and returns all the side effects
+func (s *SellOrderBook) FillBuyOrder(order Order) (
+	remainingBuyOrder Order,
+	liquidated []Order,
+	purchase int32,
+	filled bool,
+) {
+	var liquidatedList []Order
+	totalPurchase := int32(0)
+	remainingBuyOrder = order
+
+	// Liquidate as long as there is match
+	for {
+		var match bool
+		var liquidation Order
+		remainingBuyOrder, liquidation, purchase, match, filled = s.liquidateFromBuyOrder(
+			remainingBuyOrder,
+		)
+		if !match {
+			break
+		}
+
+		// Update gains
+		totalPurchase += purchase
+
+		// Update liquidated
+		liquidatedList = append(liquidatedList, liquidation)
+
+		if filled {
+			break
+		}
+	}
+
+	return remainingBuyOrder, liquidatedList, totalPurchase, filled
+}
+
+// liquidateFromBuyOrder liquidates the first sell order of the book from the buy order
+// if no match is found, return false for match
+func (s *SellOrderBook) liquidateFromBuyOrder(order Order) (
+	remainingBuyOrder Order,
+	liquidatedSellOrder Order,
+	purchase int32,
+	match bool,
+	filled bool,
+) {
+	remainingBuyOrder = order
+
+	// No match if no order
+	orderCount := len(s.Book.Orders)
+	if orderCount == 0 {
+		return order, liquidatedSellOrder, purchase, false, false
+	}
+
+	// Check if match
+	lowestAsk := s.Book.Orders[orderCount-1]
+	if order.Price < lowestAsk.Price {
+		return order, liquidatedSellOrder, purchase, false, false
+	}
+
+	liquidatedSellOrder = *lowestAsk
+
+	// Check if buy order can be entirely filled
+	if lowestAsk.Amount >= order.Amount {
+		remainingBuyOrder.Amount = 0
+		liquidatedSellOrder.Amount = order.Amount
+		purchase = order.Amount
+
+		// Remove lowest ask if it has been entirely liquidated
+		lowestAsk.Amount -= order.Amount
+		if lowestAsk.Amount == 0 {
+			s.Book.Orders = s.Book.Orders[:orderCount-1]
+		} else {
+			s.Book.Orders[orderCount-1] = lowestAsk
+		}
+
+		return remainingBuyOrder, liquidatedSellOrder, purchase, true, true
+	}
+
+	// Not entirely filled
+	purchase = lowestAsk.Amount
+	s.Book.Orders = s.Book.Orders[:orderCount-1]
+	remainingBuyOrder.Amount -= lowestAsk.Amount
+
+	return remainingBuyOrder, liquidatedSellOrder, purchase, true, false
+}
